Done stage 9, refact sss_core/lib
