+++ programs/sss_core/src/instructions/mod.rs
@@ -1,0 +1,9 @@
+pub mod initialize;
+pub mod mint;
+pub mod burn;
+pub mod seize;
+
+pub use initialize::*;
+pub use mint::*;
+pub use burn::*;
+pub use seize::*;
